+++ CHANGES.rst
@@ -4,7 +4,11 @@ CHANGES
 `Unreleased <https://github.com/fmigneault/aiu/tree/master>`_ (latest)
 ------------------------------------------------------------------------------------
 
-* Nothing yet.
+* Add multiple heuristic rules to attempt matching ambiguous file names against provided audio information.
+* Add heuristics and patched characters conditions to better detect ambiguous file names renamed following download
+  to better detect them once again on subsequent download operation, taking advantage of cached file contents.
+* Add CLI options to allow toggling of experimental heuristics that are more prone to errors than typical "hard"
+  matching conditions, at the expense of potential failure to resolve more complicated matching cases.
 
 `1.8.0 <https://github.com/fmigneault/aiu/tree/1.8.0>`_ (2022-09-03)
 ------------------------------------------------------------------------------------